[Refactor] #4 - Refactoring File Tree
