+++ src/pages/main/_components/boothList/MainBoothList.styled.ts
