+++ src/pages/main/_components/navigation/MainNavigation.styled.ts
